+++ app/src/main/java/br/com/qmovie/service/TmdbRepository.kt
@@ -1,4 +1,4 @@
-package br.com.qmovie
+package br.com.qmovie.service
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
